requirements.txt
# ─────────────────────────────────────────────
# fraud-shield-ml | requirements.txt
# Cross-platform: macOS / Windows / Linux
# Python >= 3.10
# ─────────────────────────────────────────────

# Core data
numpy>=1.26.0
pandas>=2.2.0
scipy>=1.13.0

# Visualisation
matplotlib>=3.8.0
seaborn>=0.13.0
plotly>=5.22.0

# Machine learning
scikit-learn>=1.5.0
xgboost>=2.0.0
lightgbm>=4.3.0

# Imbalance handling
imbalanced-learn>=0.12.0

# Hyperparameter tuning
optuna>=3.6.0
optuna-integration>=3.6.0

# Explainability
shap>=0.45.0
lime>=0.2.0.1

# Reporting / PDF
reportlab>=4.2.0
Pillow>=10.3.0
jinja2>=3.1.0
weasyprint>=62.0

# Streamlit dashboard
streamlit>=1.35.0
streamlit-option-menu>=0.3.12

# Utilities
joblib>=1.4.0
tqdm>=4.66.0
python-dotenv>=1.0.0
loguru>=0.7.0
pyyaml>=6.0.1

# Testing
pytest>=8.2.0
pytest-cov>=5.0.0
pyarrow>=7.0
